requirement.txt

geopy 
django-leaflet
